change the type

+++ 0x14-bit_manipulation/1-print_binary.c
@@ -7,7 +7,7 @@
  */
 void print_binary(unsigned long int n)
 {
-	int bits = 31;
+	unsigned long int bits = 31;
 	int flag = 0;
 
 	if (n == 0)
